Promote CowData to 64 bits
Fixes a lot of bugs, please help me fill the list.

(cherry picked from commit 24fe0715b53c6e599adfea9f69716db279be9db4)

+++ modules/gdscript/gdscript_compiler.cpp
@@ -1371,7 +1371,7 @@ GDScriptCodeGenerator::Address GDScriptCompiler::_parse_expression(CodeGen &code
 				return GDScriptCodeGenerator::Address();
 			}
 
-			main_script->lambda_info.insert(function, { lambda->captures.size(), lambda->use_self });
+			main_script->lambda_info.insert(function, { (int)lambda->captures.size(), lambda->use_self });
 			gen->write_lambda(result, function, captures, lambda->use_self);
 
 			for (int i = 0; i < captures.size(); i++) {
